requirements.txt
gunicorn
django-heroku
Pillow
django-crispy-forms
python-dotenv
requests
sodapy
python-dateutil
black
flake8
coverage
coveralls
re